SCRUM, Ramble Plan, and KPI Dashboard
I created the SCRUM board and populated it with numerous (hopefully all of the) user stories I will need to make the program.

I wrote out a ramble plan just to get my ideas down. There are various aspects of this plan that have been abandoned, and the SCRUM board is the better source of information.

I wrote the KPI Dashboard file as just a way to explain to myself (and anyone else) what the KPI Dashboard is and what each goal and sub-goal are.