A lot of changes I've been sitting on.
I've been working in a separate folder. It started as just testing out certain features, but then I found it was easier to code those features to completion by themselves and then insert them here. As such, these are some changes I made a while ago that I'm pushing because I'm about to import a lot of code that I've written.