PDF Rendering.
This is going to be changed slightly, as I make sure that each report has it's own HTML and CSS saved with the report. So the defaults will be contained within the `defaults.py` file, and saved into the home_directory, and onloaded into each year report when the year is created.